Refactor to move api and config wrapper code to an sa_util module, shared across applications
